% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadModel.R
\name{loadSmokingModel}
\alias{loadSmokingModel}
\title{Load the smoking model}
\usage{
loadSmokingModel()
}
\value{
Returns a plpModel obect that can be used to apply the smoking model
to new plpData via the function PatientLevelPrediction::predictPlp
}
\description{
This function loads the smoking model from the package
}
\details{
Use this function to load a plpModel that predicts whether a patient is currently a smoker or non-smoker
the value returned between 0 and 1 is the probability of being a smoker (1 = 100% and 0 = 0%).
}
