% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{viewSmokingCoefficients}
\alias{viewSmokingCoefficients}
\title{View Smoking prediction model coefficients}
\usage{
viewSmokingCoefficients(packageName = "SmokingModel")
}
\description{
View Smoking prediction model coefficients
}
\details{
This will show a table with the covaraite name and coefficient value
}
