% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{viewSmokingShiny}
\alias{viewSmokingShiny}
\title{View Smoking risk prediction model in shiny app}
\usage{
viewSmokingShiny(packageName = "SmokingModel")
}
\description{
View Smoking risk prediction model in shiny app
}
\details{
This will open a shiny app to explore the Smoking model
}
